+++ backend/src/migrations.rs
@@ -280,7 +280,9 @@ pub enum SchemaVersion {
     CreatePltTokenAndEventTables,
     #[display("0040: Alter PLT events add event_timestamp and index")]
     AlterPltEventsAddEventTimestampAndIndex,
-    #[display("0041: Alter PLT accounts add NOT NULL constraint on amount and decimal")]
+    #[display("0041: Re-Add the index `baker_related_tx_idx` that was dropped in mistake due to column drop")]
+    ReAddBakerRelatedTransactionsIndex,
+    #[display("0042: Alter PLT accounts add NOT NULL constraint on amount and decimal")]
     AlterPltAccountsAddNotNullConstraint,
 }
 impl SchemaVersion {
@@ -352,6 +354,7 @@ impl SchemaVersion {
             SchemaVersion::BakerApyQueryUpdateProtectAgainstOverflow => false,
             SchemaVersion::CreatePltTokenAndEventTables => false,
             SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => false,
+            SchemaVersion::ReAddBakerRelatedTransactionsIndex => false,
             SchemaVersion::AlterPltAccountsAddNotNullConstraint => false,
         }
     }
@@ -403,6 +406,7 @@ impl SchemaVersion {
             SchemaVersion::BakerApyQueryUpdateProtectAgainstOverflow => false,
             SchemaVersion::CreatePltTokenAndEventTables => false,
             SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => false,
+            SchemaVersion::ReAddBakerRelatedTransactionsIndex => false,
             SchemaVersion::AlterPltAccountsAddNotNullConstraint => false,
         }
     }
@@ -681,7 +685,15 @@ impl SchemaVersion {
             SchemaVersion::AlterPltEventsAddEventTimestampAndIndex => {
                 tx.as_mut()
                     .execute(sqlx::raw_sql(include_str!(
-                        "./migrations/m0041-plt-account-alter-add-not-null-constraint.sql"
+                        "./migrations/m0041_re_add_index_baker_related_transactions.sql"
+                    )))
+                    .await?;
+                SchemaVersion::ReAddBakerRelatedTransactionsIndex
+            }
+            SchemaVersion::ReAddBakerRelatedTransactionsIndex => {
+                tx.as_mut()
+                    .execute(sqlx::raw_sql(include_str!(
+                        "./migrations/m0042-plt-account-alter-add-not-null-constraint.sql"
                     )))
                     .await?;
                 SchemaVersion::AlterPltAccountsAddNotNullConstraint